This test appears to have failed just waiting too short a time for a snapshot to be made.
Now that there is a stable name for hourly snapshots (hourly.X),
add a function that can wait a specified number of seconds for that
file to arrive, and return with the existence of the file.

Rewrite the test without sleep $SLEEP.  Should be faster in most
cases, and slower when it needs to be.  Up to 30 seconds is permitted
for a snapshot.